tests: Simplify tests for no cms db calls
