+++ tests/Dfe.PlanTech.Application.UnitTests/Content/Queries/GetSubTopicRecommendationQueryTests.cs
@@ -1,8 +1,11 @@
 using Dfe.PlanTech.Application.Caching.Interfaces;
 using Dfe.PlanTech.Application.Content.Queries;
+using Dfe.PlanTech.Application.Persistence.Interfaces;
+using Dfe.PlanTech.Application.Persistence.Models;
 using Dfe.PlanTech.Domain.Content.Models;
-using Dfe.PlanTech.Domain.Content.Queries;
+using Dfe.PlanTech.Domain.Questionnaire.Enums;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
+using Dfe.PlanTech.Infrastructure.Application.Models;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 
@@ -10,106 +13,129 @@ namespace Dfe.PlanTech.Application.UnitTests.Content.Queries;
 
 public class GetSubTopicRecommendationQueryTests
 {
+    private readonly IContentRepository _repoSubstitute = Substitute.For<IContentRepository>();
     private readonly GetSubTopicRecommendationQuery _query;
-    private readonly ILogger<GetSubTopicRecommendationQuery> _logger = Substitute.For<ILogger<GetSubTopicRecommendationQuery>>();
-    private readonly IGetSubTopicRecommendationQuery _contentfulQuery = Substitute.For<IGetSubTopicRecommendationQuery>();
-    private readonly IGetSubTopicRecommendationQuery _dbQuery = Substitute.For<IGetSubTopicRecommendationQuery>();
-    private readonly ICmsCache _cache = Substitute.For<ICmsCache>();
-    private readonly SubtopicRecommendation _subtopicRecommendation;
+    private readonly Section _subTopicOne;
+    private readonly Section _subTopicTwo;
+
+    private readonly SubtopicRecommendation? _subtopicRecommendationOne;
+    private readonly SubtopicRecommendation? _subtopicRecommendationTwo;
+
     private readonly List<SubtopicRecommendation> _subtopicRecommendations = [];
+    private readonly ILogger<GetSubTopicRecommendationQuery> _logger = Substitute.For<ILogger<GetSubTopicRecommendationQuery>>();
+    private readonly ICmsCache _cache = Substitute.For<ICmsCache>();
 
     public GetSubTopicRecommendationQueryTests()
     {
         var recommendationSectionOne = new RecommendationSection()
         {
-            Answers =
-            [
-                new(),
-                new(),
-                new()
-            ],
-            Chunks =
-            [
-                new()
+            Answers = new List<Answer>()
+            {
+                new Answer(),
+                new Answer(),
+                new Answer()
+            },
+            Chunks = new List<RecommendationChunk>()
+            {
+                new RecommendationChunk()
                 {
-                    Answers =
-                    [
-                        new()
+                    Answers = new List<Answer>()
+                    {
+                        new Answer()
                         {
                             Sys = new SystemDetails() { Id = "1" }
                         },
-                        new()
+                        new Answer()
                         {
                             Sys = new SystemDetails() { Id = "2" }
                         },
-                        new()
+                        new Answer()
                         {
                             Sys = new SystemDetails() { Id = "3" }
                         }
-                    ],
+                    },
                     Header = "chunk1"
-                },
-                new()
-                {
-                    Answers =
-                    [
-                        new()
-                        {
-                            Sys = new SystemDetails() { Id = "4" }
-                        },
-                        new()
-                        {
-                            Sys = new SystemDetails() { Id = "5" }
-                        },
-                        new()
-                        {
-                            Sys = new SystemDetails() { Id = "6" }
-                        }
-                    ],
-                    Header = "chunk3",
+
                 },
                 new RecommendationChunk()
                 {
-                    Answers =
-                    [
-                        new()
+                    Answers = new List<Answer>()
+                    {
+                        new Answer()
                         {
-                            Sys = new() { Id = "7" }
+                            Sys = new SystemDetails() { Id = "4" }
                         },
-                        new()
+                        new Answer()
                         {
-                            Sys = new() { Id = "8" }
+                            Sys = new SystemDetails() { Id = "5" }
                         },
-                        new()
+                        new Answer()
                         {
-                            Sys = new() { Id = "9" }
+                            Sys = new SystemDetails() { Id = "6" }
                         }
-                    ],
+                    },
+                    Header = "chunk3"
+
+                },
+                new RecommendationChunk()
+                {
+                    Answers = new List<Answer>()
+                    {
+                        new Answer()
+                        {
+                            Sys = new SystemDetails() { Id = "7" }
+                        },
+                        new Answer()
+                        {
+                            Sys = new SystemDetails() { Id = "8" }
+                        },
+                        new Answer()
+                        {
+                            Sys = new SystemDetails() { Id = "9" }
+                        }
+                    },
                     Header = "chunk2"
                 }
-            ]
-        };
-
-        var subtopicOne = new Section() { Sys = new() { Id = "SubTopicId" } };
-
-        _subtopicRecommendation = new()
-        {
-            Intros =
-            [
-                new() { Maturity = "Low" },
-                new() { Maturity = "Medium" },
-                new() { Maturity = "High" },
-            ],
-            Section = recommendationSectionOne,
-            Subtopic = subtopicOne,
-            Sys = new()
-            {
-                Id = "Subtopic-Recommendation-Id"
             }
         };
 
-        _subtopicRecommendations.Add(_subtopicRecommendation);
-        _query = new GetSubTopicRecommendationQuery(_contentfulQuery, _dbQuery, _logger, _cache);
+        _subTopicOne = new()
+        {
+            Sys = new SystemDetails()
+            {
+                Id = "SubTopicId"
+            }
+        };
+
+        _subTopicTwo = new()
+        {
+            Sys = new SystemDetails()
+            {
+                Id = "IdForAnotherSubTopic"
+            }
+        };
+
+        _subtopicRecommendationOne = new()
+        {
+            Intros =
+            [
+                new RecommendationIntro() { Maturity = "Low", Header = new Header() { Text = "Low-Maturity-Intro" } },
+                new RecommendationIntro() { Maturity = "Medium", Header = new Header() { Text = "Medium-Maturity-Intro" } },
+                new RecommendationIntro() { Maturity = "High", Header = new Header() { Text = "High-Maturity-Intro" } },
+            ],
+            Section = recommendationSectionOne,
+            Subtopic = _subTopicOne,
+        };
+
+        _subtopicRecommendationTwo = new()
+        {
+            Subtopic = _subTopicTwo
+        };
+
+        _subtopicRecommendations.Add(_subtopicRecommendationOne);
+        _subtopicRecommendations.Add(_subtopicRecommendationTwo);
+
+        _query = new(_repoSubstitute, _logger, _cache);
         _cache.GetOrCreateAsync(Arg.Any<string>(), Arg.Any<Func<Task<SubtopicRecommendation>>>())
             .Returns(callInfo =>
             {
@@ -119,54 +145,130 @@ public class GetSubTopicRecommendationQueryTests
     }
 
     [Fact]
-    public async Task Returns_Result_From_Contentful()
+    public async Task GetSubTopicRecommendation_Returns_Correct_SubTopicRecommendation_From_SectionOne()
     {
-        _contentfulQuery.GetSubTopicRecommendation(Arg.Any<string>(), Arg.Any<CancellationToken>())
-                .Returns((callinfo) =>
-                {
-                    var subtopicId = _subtopicRecommendation.Subtopic.Sys.Id;
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([_subtopicRecommendationOne]);
 
-                    return _subtopicRecommendations.FirstOrDefault(recommendation => recommendation.Subtopic.Sys.Id == subtopicId);
-                });
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation(_subTopicOne.Sys.Id);
+
+        Assert.NotNull(subTopicRecommendation);
+        Assert.Equal(_subtopicRecommendationOne, subTopicRecommendation);
+    }
 
 
-        var subtopicRecommendation = await _query.GetSubTopicRecommendation(_subtopicRecommendation.Subtopic.Sys.Id, CancellationToken.None);
 
-        Assert.NotNull(subtopicRecommendation);
-        Assert.Equal(subtopicRecommendation.Sys.Id, _subtopicRecommendation!.Sys.Id);
-        Assert.Equal(subtopicRecommendation.Subtopic.Sys.Id, _subtopicRecommendation!.Subtopic.Sys.Id);
+    [Fact]
+    public async Task GetSubTopicRecommendation_Returns_Null_When_Not_Found()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([null]);
 
-        Assert.Single(_logger.GetMatchingReceivedMessages(
-            $"Retrieved subtopic recommendation {_subtopicRecommendation.Subtopic.Sys.Id} from Contentful", LogLevel.Trace));
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation("not a real id");
+
+        Assert.Null(subTopicRecommendation);
+        Assert.Single(_logger.ReceivedCalls());
     }
 
     [Fact]
-    public async Task Returns_Null_When_Not_Found()
+    public async Task GetSubTopicRecommendation_Returns_Correct_SubTopicRecommendation_From_SectionTwo()
     {
-        _contentfulQuery.GetSubTopicRecommendation(Arg.Any<string>(), Arg.Any<CancellationToken>())
-                .Returns(ReturnNullSubtopicRecommendation());
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([_subtopicRecommendationTwo]);
 
-        var result = await _query.GetSubTopicRecommendation(_subtopicRecommendation.Subtopic.Sys.Id, CancellationToken.None);
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation(_subTopicTwo.Sys.Id);
+
+        Assert.NotNull(subTopicRecommendation);
+        Assert.Equal(_subtopicRecommendationTwo, subTopicRecommendation);
+    }
+
+    [Fact]
+    public async Task GetSubTopicRecommendationIntro_Returns_Intro_When_Exists_In_SubTopicRecommendation_From_SectionOne()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([_subtopicRecommendationOne]);
+
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation(_subTopicOne.Sys.Id);
+
+        var intro = subTopicRecommendation!.GetRecommendationByMaturity(Maturity.Medium.ToString());
+
+        Assert.NotNull(intro);
+        Assert.Equal(Maturity.Medium.ToString(), intro.Maturity);
+    }
+
+    [Fact]
+    public async Task GetSubTopicRecommendationChunk_When_List_Of_Answers_Passed_To_RecommendationSection_From_SectionOne()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([_subtopicRecommendationOne]);
+
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation(_subTopicOne.Sys.Id);
+
+        var chunks = subTopicRecommendation!.Section.GetRecommendationChunksByAnswerIds(new List<string>() { "1", "5", "9" });
+
+        Assert.Equal(3, chunks.Count);
+    }
+
+    [Fact]
+    public async Task GetSubTopicRecommendationChunk_When_List_Of_Answers_Passed_To_RecommendationSection_From_SectionOne_No_Duplicate_Chunks_Returned()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([_subtopicRecommendationOne]);
+
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation(_subTopicOne.Sys.Id);
+
+        var chunks = subTopicRecommendation!.Section.GetRecommendationChunksByAnswerIds(["1", "7", "9"]);
+
+        Assert.Equal(2, chunks.Count);
+    }
+
+    [Fact]
+    public async Task GetSubTopicRecommendationChunks_Returns_Empty_When_List_Of_Answers_Passed_To_RecommendationSection_From_SectionOne_Does_Not_Have_Any_Chunks_Associated()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation?>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>()).Returns([_subtopicRecommendationOne]);
+
+        SubtopicRecommendation? subTopicRecommendation = await _query.GetSubTopicRecommendation(_subTopicOne.Sys.Id);
+
+        var chunks = subTopicRecommendation!.Section.GetRecommendationChunksByAnswerIds(["10"]);
+
+        Assert.Empty(chunks);
+    }
+
+    [Fact]
+    public async Task GetRecommendationsViewDto_Retrieves_Correct_Information_When_Existing()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>())
+                        .Returns((callinfo) =>
+                        {
+                            var options = callinfo.ArgAt<GetEntitiesOptions>(0);
+                            var idFilter = options.Queries!.Where(query => query.Field == "fields.subtopic.sys.id")
+                                                            .Select(query => query as ContentQueryEquals)
+                                                            .Select(query => query!.Value).FirstOrDefault();
+
+                            return _subtopicRecommendations.Where(rec => rec.Subtopic.Sys.Id == idFilter)
+                                                            .Select(rec => new SubtopicRecommendation()
+                                                            {
+                                                                Intros = rec.Intros,
+                                                                Sys = rec.Sys
+                                                            });
+                        });
+
+        var maturity = "Low";
+        var expectedIntro = _subtopicRecommendationOne!.Intros.FirstOrDefault(intro => intro.Maturity == maturity);
+
+        Assert.NotNull(expectedIntro);
+
+        var result = await _query.GetRecommendationsViewDto(_subtopicRecommendationOne!.Subtopic.Sys.Id, maturity);
+
+        Assert.NotNull(result);
+        Assert.Equal(result.RecommendationSlug, expectedIntro.Slug);
+        Assert.Equal(result.DisplayName, expectedIntro.HeaderText);
+    }
+
+    [Fact]
+    public async Task GetRecommendationsViewDto_Returns_Null_When_No_Match()
+    {
+        _repoSubstitute.GetEntities<SubtopicRecommendation>(Arg.Any<GetEntitiesOptions>(), Arg.Any<CancellationToken>())
+                        .Returns((callinfo) => []);
+
+        var result = await _query.GetRecommendationsViewDto("any id", "Low");
 
         Assert.Null(result);
 
-        AssertFailureMessageLog(_subtopicRecommendation);
-
-        static Func<NSubstitute.Core.CallInfo, SubtopicRecommendation?> ReturnNullSubtopicRecommendation()
-        {
-            return (callinfo) =>
-            {
-                SubtopicRecommendation? recommendation = null;
-
-                return recommendation;
-            };
-        }
-    }
-
-    private void AssertFailureMessageLog(SubtopicRecommendation? subtopicRecommendation)
-    {
-        Assert.Single(_logger.GetMatchingReceivedMessages(
-            $"Was unable to find a subtopic recommendation for {subtopicRecommendation?.Subtopic.Sys.Id} from DB or Contentful",
-            LogLevel.Error));
+        Assert.Single(_logger.GetMatchingReceivedMessages("Could not find subtopic recommendation in Contentful for any id", LogLevel.Error));
     }
 }
